issue #5: Add function to filter the investment files
The new function allows filtering the collected data files according to a list of dates or the latest date.
Thus, after collecting the data files, this function is called to apply the argument requirements or process
all the available data or just the latest one. The idea is to use in the following the same code for both
cases without checking again the input argument.